+++ src/systems/subspace/modules/session/package.json
@@ -26,7 +26,6 @@
     "@metorial-subspace/module-deployment": "^1.0.0",
     "@metorial-subspace/module-provider-internal": "^1.0.0",
     "@metorial-subspace/module-tenant": "^1.0.0",
-    "@metorial-subspace/presenters": "^1.0.0",
     "p-queue": "^8.0.1"
   },
   "devDependencies": {
