+++ src/systems/subspace/modules/session/src/index.ts
@@ -1,3 +1,4 @@
 export * from './services';
 export * from './lib/adminProviderTelemetryErrorGroups';
+export * from './lib/providerTelemetryErrorGroupExport';
 export * from './lib/sessionProviderNameTemplate';
